+++ styx-e2e-test/src/test/resources/styx-e2e-test-scheduler.conf
@@ -13,6 +13,8 @@ styx.gke.default.namespace = ${styx.test.namespace}
 styx.flyte.enabled = false # TODO: Make end to end test for flyte
 styx.flyte.admin.host = ""
 styx.flyte.admin.port = 1234
+styx.flyte.admin.insecure = true
+
 
 # k8s request timeout in ms
 styx.k8s.request-timeout = 60000
