+++ styx-scheduler-service/src/test/java/com/spotify/styx/state/handlers/FlyteRunnerHandlerTest.java
@@ -39,6 +39,7 @@ import com.spotify.styx.state.EventRouter;
 import com.spotify.styx.state.RunState;
 import com.spotify.styx.state.RunState.State;
 import com.spotify.styx.state.StateData;
+import com.spotify.styx.util.IsClosedException;
 import java.util.Optional;
 import java.util.function.Function;
 import junitparams.JUnitParamsRunner;
