+++ styx-scheduler-service/src/test/java/com/spotify/styx/state/handlers/StyxIdToFlyteExecNameMapperTest.java
@@ -28,7 +28,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
 
-import com.spotify.styx.state.handlers.FlyteRunnerHandler.StyxIdToFlyteExecNameMapper;
 import java.util.stream.Stream;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
